requirements.txt
flask
flask-cors
flask_sqlalchemy
flask_login
dotenv
torch
transformer
openai
plyer
geocoder
PyPDF2